Moved test running logic to own file
